Improve p_FunnyShape member layout
