+++ include/ffcc/p_FunnyShape.h
@@ -7,6 +7,46 @@
 #include "ffcc/system.h"
 #include "ffcc/FS_USB_Process.h"
 
+#ifdef FFCC_P_FUNNYSHAPE_REAL_PTRARRAY
+struct _GXTexObj;
+
+class CUSBStreamData
+{
+public:
+    CUSBStreamData();
+    ~CUSBStreamData();
+
+    void CreateBuffer();
+    void DeleteBuffer();
+    int IsUSBStreamDataDone();
+    void SetUSBStreamDataDone();
+
+    u8* m_data;
+    int m_headerReady;
+    int m_dataReady;
+    u32 m_sizeBytes;
+    int m_packetCode;
+};
+
+template <class T>
+class CPtrArray
+{
+public:
+    virtual ~CPtrArray();
+
+    unsigned long size;
+    unsigned long numItems;
+    unsigned long defaultSize;
+    T* items;
+    CMemory::CStage* stage;
+    int growCapacity;
+
+    CPtrArray();
+
+    void RemoveAll();
+    void DeleteAndRemoveAll();
+};
+#else
 struct CUSBStreamDataStorage
 {
     CUSBStreamDataStorage();
@@ -14,6 +54,7 @@ struct CUSBStreamDataStorage
 
     u8 m_storage[0x14];
 };
+#endif
 
 class CFunnyShapePcs : public CProcess
 {
@@ -41,12 +82,21 @@ public:
 
     CMemory::CStage* m_viewerStage;        // 0x04
     u8 m_viewerState[0x34];                // 0x08
+#ifdef FFCC_P_FUNNYSHAPE_REAL_PTRARRAY
+    CUSBStreamData m_usbStreamDataStorage; // 0x3C
+#else
     CUSBStreamDataStorage m_usbStreamDataStorage; // 0x3C
+#endif
     CFunnyShape m_funnyShape;              // 0x50
     FS_DISPLAY_STATUS m_displayPending;    // 0x6178
     u32 m_displayTextureEnabled;           // 0x61B8
+#ifdef FFCC_P_FUNNYSHAPE_REAL_PTRARRAY
+    CPtrArray<OSFS_TEXTURE_ST*> m_textureHeaders; // 0x61BC
+    CPtrArray<_GXTexObj*> m_textureObjects;       // 0x61D8
+#else
     u8 m_texturePtrArrayStorage[0x1C];     // 0x61BC
     u8 m_gxTexObjPtrArrayStorage[0x1C];    // 0x61D8
+#endif
 };
 
 #ifdef FFCC_DEFINE_FUNNYSHAPEPCS_STORAGE
